+++ utilities/pytest_utils.py
@@ -1,10 +1,6 @@
 import shutil
 from typing import Optional
 
-from simple_logger.logger import get_logger
-
-LOGGER = get_logger(name=__name__)
-
 
 def separator(symbol_: str, val: Optional[str] = None) -> str:
     terminal_width = shutil.get_terminal_size(fallback=(120, 40))[0]
